drop old file
