+++ nodes/svg/svg_document.py
@@ -26,7 +26,7 @@ from bpy.props import (
 import sverchok
 from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.data_structure import updateNode
-from sverchok.utils.sv_operator_utils import SvGenericNodeLocator
+from sverchok.utils.sv_operator_mixins import SvGenericNodeLocator
 
 evolver_mem = {}
