+++ utils/sv_obj_baker.py
@@ -11,7 +11,7 @@ from bpy.props import StringProperty
 from mathutils import Vector, Matrix
 
 from sverchok.data_structure import node_id, dataCorrect, dataCorrect_np
-from sverchok.utils.sv_operator_utils import SvGenericNodeLocator
+from sverchok.utils.sv_operator_mixins import SvGenericNodeLocator
 
 cache_viewer_baker = {}
