+++ utils/sv_operator_mixins.py
@@ -1,22 +1,40 @@
-# ##### BEGIN GPL LICENSE BLOCK #####
-#
-#  This program is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU General Public License
-#  as published by the Free Software Foundation; either version 2
-#  of the License, or (at your option) any later version.
-#
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with this program; if not, write to the Free Software Foundation,
-#  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-#
-# ##### END GPL LICENSE BLOCK #####
+# This file is part of project Sverchok. It's copyrighted by the contributors
+# recorded in the version control history of the file, available from
+# its original location https://github.com/nortikin/sverchok/commit/master
+#  
+# SPDX-License-Identifier: GPL3
+# License-Filename: LICENSE
+
 
 import bpy
+from bpy.props import StringProperty
+
+class SvGenericNodeLocator():
+    """
+    usage:
+    add this to an Operator class definition if you need to track the origin
+    of a click.
+
+    see at the bottom of this file in " class SvGenericCallbackOldOp " how to use it.
+
+    """
+    tree_name: StringProperty(default='', description="name of the node tree")
+    node_name: StringProperty(default='', description="name of the node")
+
+    def get_node(self, context):
+        """ context.node is usually provided, else tree_name/node_name must be passed """
+        if self.tree_name and self.node_name:
+            return bpy.data.node_groups[self.tree_name].nodes[self.node_name]
+
+        if hasattr(context, "node"):
+            return context.node
+
+        print("treename or nodename not supplied, node not found in available trees")
+        print(f"received tree_name: {tree_name} and node_name: {node_name}")
+        return None
+
+    def get_tree(self):
+        return bpy.data.node_groups.get(self.tree_name)
 
 
 class SvGenericCallbackWithParams():
@@ -159,3 +177,36 @@ class SvGenericDirectorySelector():
         wm = context.window_manager
         wm.fileselect_add(self)
         return {'RUNNING_MODAL'}
+
+
+class SvGenericCallbackOldOp(bpy.types.Operator, SvGenericNodeLocator):
+    """ 
+    This operator is generic and will call .fn_name on the instance of the caller node
+    """
+    bl_idname = "node.sverchok_generic_callback_old"
+    bl_label = "Sverchok text input"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    fn_name: StringProperty(name='function name')
+
+    def execute(self, context):
+        n = self.get_node(context)
+        if not n:
+            return {'CANCELLED'}
+
+        f = getattr(n, self.fn_name, None)
+        if not f:
+            msg = "{0} has no function named '{1}'".format(n.name, self.fn_name)
+            self.report({"WARNING"}, msg)
+            return {'CANCELLED'}
+        f()
+
+        return {'FINISHED'}
+
+
+def register():
+    bpy.utils.register_class(SvGenericCallbackOldOp)
+
+
+def unregister():
+    bpy.utils.unregister_class(SvGenericCallbackOldOp)
